commit #2
in the nv on line 5 Test class for CurrencyConverter is added on line 6 and 7 Added @BeforeEach setup method for initializing test conditions ( currently empty) on line 9and 10 Implemented test method to verify currency conversion and  Called CurrencyConverter.convert() to check USD to INR conversion. on line 11 Asserted that conversion defaults to 1:1 rate for unsupported currency pairs.